Add k-fold validations.
