+++ .gitignore
@@ -166,3 +166,4 @@ cython_debug/
 source/
 dataset/
 data/
+results/
