+++ dataset.py
@@ -2,30 +2,35 @@ import pandas as pd
 import numpy as np
 import torch
 from torch.utils.data import Dataset
-from package import transforms
-from package.misalignment_overlap import RandomMisalignmentOverlap
+from utils import transforms
+from utils.misalignment_overlap import RandomMisalignmentOverlap
 
 transform_train = transforms.Compose([
     # transforms.RandomCrop(32, padding=4),
     transforms.RandomHorizontalFlip(),
     transforms.RandomYFlip(),
     transforms.RandomZFlip(),
-    RandomMisalignmentOverlap(),
     # transforms.ZeroOut(4),
+    transforms.ToTensor(),
+    transforms.Normalize(),
+    RandomMisalignmentOverlap(alpha=0.7),
+])
+
+transform_val = transforms.Compose([
     transforms.ToTensor(),
     transforms.Normalize(),
 ])
 
 
 class LIDCIDRIDataset(Dataset):
-    def __init__(self, root_dir, transform=transform_train):
+    def __init__(self, root_dir, is_train=True):
         sr = pd.read_csv('dataset/sr.csv')
-
         self.root_dir = root_dir
-        self.transform = transform
+        self.transform = transform_train if is_train else transform_val
         self.sr = sr
         self.sequence = np.load(f'{root_dir}/balanced_seg_ids.npy')
         self.total_data_num = len(self.sequence)
+        self.is_train = is_train
 
     def __len__(self):
         return self.total_data_num
@@ -41,4 +46,3 @@ class LIDCIDRIDataset(Dataset):
         label = row['malignancy'].values[0]
 
         return input, torch.FloatTensor(features), label
-
